# docs/_generated/studio_openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkContributeRequest": {
        "description": "Request body for benchmark-databank contribution uploads.",
        "properties": {
          "handle": {
            "default": "",
            "maxLength": 40,
            "title": "Handle",
            "type": "string"
          },
          "submission": {
            "additionalProperties": true,
            "title": "Submission",
            "type": "object"
          }
        },
        "required": [
          "submission"
        ],
        "title": "BenchmarkContributeRequest",
        "type": "object"
      },
      "BenchmarkRunRequest": {
        "description": "Request body for local Studio benchmark runs.",
        "properties": {
          "n_channels": {
            "default": 512,
            "maximum": 8192.0,
            "minimum": 16.0,
            "title": "N Channels",
            "type": "integer"
          },
          "n_taps": {
            "default": 32,
            "maximum": 256.0,
            "minimum": 4.0,
            "title": "N Taps",
            "type": "integer"
          },
          "repeats": {
            "default": 12,
            "maximum": 50.0,
            "minimum": 3.0,
            "title": "Repeats",
            "type": "integer"
          }
        },
        "title": "BenchmarkRunRequest",
        "type": "object"
      },
...
      "CompileRequest": {
        "description": "Request body for equation-to-SystemVerilog compilation.",
        "properties": {
          "equations": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Equations",
            "type": "array"
          },
          "init": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init"
          },
          "module_name": {
            "default": "sc_neuron",
            "title": "Module Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "reset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reset"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "equations"
        ],
        "title": "CompileRequest",
        "type": "object"
      },
      "DclsEvaluateRequest": {
        "description": "Request body for DCLS kernel parity evaluation.",
        "properties": {
          "centre_q88": {
            "default": 512,
            "maximum": 65535.0,
            "minimum": 0.0,
            "title": "Centre Q88",
            "type": "integer"
          },
          "n_taps": {
            "default": 8,
            "maximum": 256.0,
            "minimum": 1.0,
            "title": "N Taps",
            "type": "integer"
          },
          "sigma_q88": {
            "default": 512,
            "exclusiveMinimum": 0.0,
            "maximum": 65535.0,
            "title": "Sigma Q88",
            "type": "integer"
          },
          "spikes": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spikes"
          },
          "weights_q88": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weights Q88"
          }
        },
        "title": "DclsEvaluateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SimulateRequest": {
        "description": "Request body for direct ODE simulation in Studio.",
        "properties": {
          "current": {
            "default": 0.0,
            "title": "Current",
            "type": "number"
          },
          "dt": {
            "default": 0.1,
            "exclusiveMinimum": 0.0,
            "title": "Dt",
            "type": "number"
          },
          "duration": {
            "default": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Duration",
            "type": "number"
          },
          "equations": {
            "items": {
              "type": "string"
            },
            "title": "Equations",
            "type": "array"
          },
          "init": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Init"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "protocol": {
            "default": "constant",
            "title": "Protocol",
            "type": "string"
          },
          "reset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reset"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "equations"
        ],
        "title": "SimulateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SC-NeuroCore Studio",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/benchmarks/contribute": {
      "post": {
        "operationId": "api_benchmark_contribute_api_benchmarks_contribute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkContributeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Benchmark Contribute Api Benchmarks Contribute Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Benchmark Contribute"
      }
    },
...
    "/api/benchmarks/run": {
      "post": {
        "operationId": "api_benchmark_run_api_benchmarks_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BenchmarkRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Benchmark Run Api Benchmarks Run Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Benchmark Run"
      }
    },
...
    "/api/compile": {
      "post": {
        "operationId": "api_compile_api_compile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Compile Api Compile Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Compile"
      }
    },
...
    "/api/dcls/evaluate": {
      "post": {
        "operationId": "api_dcls_evaluate_api_dcls_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DclsEvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Dcls Evaluate Api Dcls Evaluate Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Dcls Evaluate"
      }
    },
...
    "/api/ir/build": {
      "post": {
        "operationId": "api_ir_build_api_ir_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Ir Build Api Ir Build Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Ir Build"
      }
    },
...
    "/api/ir/emit-sv": {
      "post": {
        "operationId": "api_ir_emit_sv_api_ir_emit_sv_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Ir Emit Sv Api Ir Emit Sv Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Ir Emit Sv"
      }
    },
    "/api/ir/emit-sv-direct": {
      "post": {
        "operationId": "api_ir_emit_sv_direct_api_ir_emit_sv_direct_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Ir Emit Sv Direct Api Ir Emit Sv Direct Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Ir Emit Sv Direct"
      }
    },
    "/api/ir/verify": {
      "post": {
        "operationId": "api_ir_verify_api_ir_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Ir Verify Api Ir Verify Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Ir Verify"
      }
    },
...
    "/api/models/{name}": {
      "get": {
        "operationId": "api_model_api_models__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Model Api Models  Name  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Model"
      }
    },
    "/api/models/{name}/doc": {
      "get": {
        "operationId": "api_model_doc_api_models__name__doc_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Model Doc Api Models  Name  Doc Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Model Doc"
      }
    },
...
    "/api/nir/compile": {
      "post": {
        "operationId": "api_nir_compile_api_nir_compile_post",
        "parameters": [
          {
            "in": "query",
            "name": "module_name",
            "required": false,
            "schema": {
              "default": "sc_nir_network",
              "title": "Module Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source_kind",
            "required": false,
            "schema": {
              "default": "lfsr",
              "title": "Source Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Nir Compile Api Nir Compile Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Nir Compile"
      }
    },
...
    "/api/templates/{name}": {
      "get": {
        "operationId": "api_template_api_templates__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Api Template Api Templates  Name  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Template"
      }
    },
...
  }
}